♻️ [refactor]: #452 Phase 2: lexer_peek.rs を分離
peek / peek_at の 8 テストを `src/lexer/tests/lexer_peek.rs` に分離。

- 293 件グリーン維持
- 短縮テスト関数名集合不変 / 重複 0
- lexer_peek.rs: 64 行 (< 800)

Refs #452
